smv vv: update qsmv and associated scripts to allow multiple instances to run on a node

+++ Utilities/Scripts/stopXserver.sh
@@ -1,10 +1,27 @@
 #!/bin/bash
+IDFILE=$1
 if [ "`uname`" != "Darwin" ]; then
-  display_port=`id -u`
-  lockfile=/tmp/startXlock$display_port
   PAUSE=1
   echo "shutting down graphics environment (pausing $PAUSE s)"
   sleep $PAUSE
-  kill $SMV_ID
-  rm -f $lockfile
+  if [ -e $IDFILE ]; then
+    SMV_ID=`head -1 $IDFILE`
+    PID=`ps -el | awk '{print $4}' | grep $SMV_ID`
+    if [ "$PID" != "" ]; then
+      if [ $PID -eq $SMV_ID ]; then
+        kill $SMV_ID
+        KILLED=1
+      fi
+    fi
+  fi
 fi
+if [ "$KILLED" != "" ]; then
+  echo process $SMV_ID killed >> $IDFILE
+else
+  if [ "$SMV_ID" != "" ]; then
+    echo process $SMV_ID is not running >> $IDFILE
+  else
+    echo process does not exist >> $IDFILE
+  fi
+fi
+
